put compare_pc_and_bpod_times to pipeline/ingest/util.py
